Cleanup: warnings

+++ source/blender/blenkernel/intern/bpath.c
@@ -595,7 +595,7 @@ void BKE_bpath_traverse_id(Main *bmain, ID *id, BPathVisitor visit_cb, const int
 						}
 						else if ((seq->type == SEQ_TYPE_IMAGE) && se) {
 							/* might want an option not to loop over all strips */
-							unsigned int len = (unsigned int)MEM_allocN_len(se) / sizeof(*se);
+							unsigned int len = (unsigned int)MEM_allocN_len(se) / (unsigned int)sizeof(*se);
 							unsigned int i;
 
 							if (flag & BKE_BPATH_TRAVERSE_SKIP_MULTIFILE) {
